Updated docs

+++ src/lexer/token.zig
@@ -2,5 +2,5 @@ const tokenType = @import("token_type.zig");
 
 pub const Token = struct {
     type: tokenType.TokenType,
-    value: ?[]const usize
+    value: ?[]const u8
 };
